server: Work around incorrect include in gnutls/socket.h
On MinGW:

In file included from crypto.c:56:
/usr/x86_64-w64-mingw32/sys-root/mingw/include/gnutls/socket.h:32:10: fatal error: sys/socket.h: No such file or directory
   32 | #include <sys/socket.h>
      |          ^~~~~~~~~~~~~~

gnutls/socket.h shouldn't include this header on MinGW.  As a
workaround until this is fixed, only include the header when we know
that the function we need is available.

We will need to revisit this fix when gnutls is fixed.

Link: https://gitlab.com/gnutls/gnutls/-/issues/1382
Updates: commit fd626688845324a4b3c387fa901d96e5d20ea634