tests/test-parallel-sh.sh: Small cleanups
Quote the inline script to make it easier to follow.  This requires us
to export $curr_fds to nbdkit, but that's fine.

Use $curr_fds != "" as a proxy for the existence of /proc/$$/fd so we
don't need to test that twice.

Print the parent /proc/$$/fd as well as the child, so if a leaked fd
happens it's easier to tell what it is.

These are just refactorings which should not affect the test.

(cherry picked from commit 053293f93acee196e6411baf6f2b8827069ae7bf)