delay: Fix bounds check when parsing delay parameters
Previously the check always evaluated to false.  This was because
'*r', '1000U' and UINT_MAX are all unsigned, so the multiplication
would simply wrap around.  Casting one parameter to be 64 bit should
cause everything to be promoted, fixing the check.

This bug was accidentally introduced in commit 58187831a4 ("delay:
Improve parsing and representation of delay times").  Prior to this
commit the code used signed ints and 1000LL.

Found by Coverity:

  Error: CONSTANT_EXPRESSION_RESULT (CWE-569): [#def46]
  nbdkit-1.30.7/filters/delay/delay.c:77: result_independent_of_operands: "*r * 1000U > 4294967295U /* 2147483647 * 2U + 1U */" is always false regardless of the values of its operands. This occurs as the logical operand of "if".
  #   75|       if (nbdkit_parse_unsigned (key, value, r) == -1)
  #   76|         return -1;
  #   77|->     if (*r * 1000U > UINT_MAX) {
  #   78|         nbdkit_error ("seconds parameter %s is too large: %s", key, value);
  #   79|         return -1;

Fixes: commit 58187831a4346b44e398f105163abac8f3dfb7f0