server: Avoid Coverity warning on volatile variable
Coverity warns that reading a volatile variable inside an assert() may
trigger a side-effect, and lose the effect when compiling with
-DNDEBUG.  In this particular case, our 'quit' variable must be marked
volatile, but reading it does not trigger a side effect, so the
Coverity warning is a false positive.  Still, we can shut it up by
hoisting the read outside the assesrt.

Fixes: 4871e496 ("Server: Add and use nbdkit_nanosleep", v1.13.9)