checkwrite: Fix bounds error
In the rare case where a client tries to trim or zero an underlying
plugin which has two or more extents in the region being zeroed, and
the first of those extents was marked as data (but contained zeroes),
the filter would get "stuck" on the first extent.  It would try to
read the first extent up to the end, and finally issue a zero length
read to the underlying plugin which caused an assertion failure:

nbdkit: backend.c:624: backend_pread: Assertion `backend_valid_range (c, offset, count)' failed.

In practice this only happened on ppc64le when testing the checkwrite
filter on top of the file plugin.  This is likely because of the
larger filesystem extent granularity making it more likely to hit this
corner case.  However eventually I came up with a synthetic test which
would hit this problem on any architecture (see link).

The fix is to check if we're at the end of the current extent and
break out of the "stuck" loop before issuing the zero-sized read, and
move to the next extent in the list.

The bug has been present since the filter was introduced in
December 2020 / nbdkit 1.24.

Fixes: commit 2d0954dd6431cfd8d82e873a5b99285b0fddbeb8
Link: https://listman.redhat.com/archives/libguestfs/2022-July/029496.html
(cherry picked from commit 16a13fa9e1f3c4cbc0fae89e364f832e1a41fb40)