tests: Hoist some EXTRA_DIST out of automake conditionals
We can fail to add some test files (test.tcl, test.lua) to the tarball
if compiling with those languages disabled, which would cause knock-on
failures when the tarball was used with the languages enabled.  We
already fixed this for Ruby etc, this commit fixes it for Tcl and Lua.

(cherry picked from commit 3b6763c82909c95431ff57c2fe9be1b98316b057)