common/utils: Fix generated windows-errors.c line endings
On macOS with either 'sed' or 'gsed' there were no line endings after
the #if lines, ie:

  #if defined(WSAEBADF    ) && defined(EBADF)    case WSAEBADF       : return EBADF;#endif

instead of the intended output:

  #if defined(WSAEBADF    ) && defined(EBADF)
      case WSAEBADF       : return EBADF;
  #endif

It's unclear exactly why this happens, but add \n to force sed to
print a newline.

On Linux this changes the output so it becomes:

  #if defined(WSAEBADF    ) && defined(EBADF)

      case WSAEBADF       : return EBADF;

  #endif

which is unfortunate but the code still works.