server/sockets.c: Save errno from EADDINUSE error
The current code for dealing with EADDRINUSE dates back to the
original nbdkit server commit and was a bit odd.  I think it was
possible for getaddrinfo to return no sockets and no error and for us
to ignore that case, which would be quite bad (nbdkit would run and
bind to no sockets).  Luckily this probably never happens.

More importantly let's remove the peculiar addr_in_use boolean and
just save the errno to the place where we need to print it.  This also
makes the next fix easier.

(cherry picked from commit 5892e14c85695a033d74d0acba9b3ee93422df1f)