common/utils: On macOS remove error about incomplete atomic CLOEXEC support
macOS has mkostemp, but lacks SOCK_CLOEXEC, pipe2 or accept4.  As far
as I can tell atomic SOCK_CLOEXEC-like behaviour is not possible on
this platform, or at least Gnulib does not implement any method and I
couldn't find anything else.

This #error is essentially a warning for developers that callers ought
to use atomic support if available.  We still want to keep the
warning, but it is not applicable on macOS.