tests/test-python-thread-model.sh: Skip test on non-parallel platforms
[This is the second attempt to fix this.  The first was
commit 51e5b0f6ab253459d64634898e121b5005e3218d which was reverted.]

On macOS this test failed because the Python thread model is:

max_thread_model=parallel
thread_model=serialize_all_requests

Python could support parallel, but because of lack of pipe2, accept4,
etc we're actually more limited.  This would be the same on some other
platforms too.  Skip the test on these platforms.