tests/test-python-thread-model.sh: Skip test on non-parallel platforms
On macOS this test failed because the Python thread model is:

max_thread_model=parallel
thread_model=serialize_all_requests

Python could support parallel, but because of lack of pipe2, accept4,
etc we're actually more limited.  This would be the same on some other
platforms too.  Skip the test on these platforms.