stats-filter: bucket request sizes and only count successful requests
(1) Group requests into buckets corresponding to the number of bits needed to represent
the request size. Keeping track of every encountered request size has turned out often to
be useless in practice since e.g. ZFS can issue a very large number of differently sizes
requests.

(2) Record only successful requests in the histogram to make the histogram consistent with
the total.

(3) Simplify the code by handling the collection of histogram data in the same structure
and function as the other statistics.

(4) Introduce an assertion when the histogram total differs from the separately calculated
total (to track down issue #7).

Some code had to be moved around to ensure that functions are defined before they are
called.