protocol: Diagnose extra use of NBD_OPT_STRUCTURED_REPLY
The NBD spec states that when dealing with redundant
NBD_OPT_STRUCTURED_REPLY, "the server MAY fail the repeated message
with NBD_REP_ERR_INVALID".  Although this is not mandatory, it is the
behavior qemu-nbd uses.  To date, it's mostly academic: existing
clients don't send redundant requests without one-off hacks.  But an
upcoming patch to add a new API nbd_opt_structured_reply() to libnbd
will make it easier to trigger the scenario, at which point diagnosing
the redundancy is a nicer default.