ci: Update .gitlab-ci.yml
The pipelines are now configured to trigger from merge requests
primarily. This avoids consuming limit CI minutes when merely
pushing branches to gitlab in forks. Pipelines still run on
pushes in the upstream repo.

Running pipelines on push events to forks requires an opt in :

   git push <remote> -o ci.variable=RUN_PIPELINE=1

To re-gain previous behaviour 'RUN_PIPELINE=1' can be set globally
in the fork repository.

To save storage quota, containers will never be created in forks
anymore and thus existing containers can be deleted in forks.