tests: web-server: Close connection after each request
curl 7.85 is more aggressive about opening multiple connections.  Our
dummy web server is single threaded and very simplistic so it cannot
handle this at all and this causes deadlocks.  A simple fix is just to
send "Connection: close" in the HTTP reply so curl closes and reopens
the connection after each request.  Inefficient, but this is only used
for light testing.

Also this adds a little more debugging.