crypto: Consistently avoid gnutls*_deinit on _init failure
As of gnutls 3.7.8, it is hit and miss as to whether it is safe to use
gnutls*_deinit if the corresponding gnutls*_init function failed.  I
completed an audit of all such functions in gnutls[1]; it is always
safe to skip the _deinit function, so it is easier to follow that rule
of thumb than to assume that an unconditional deinit is safe (which in
some cases is not until the next gnutls release).

[1] https://gitlab.com/gnutls/gnutls/-/merge_requests/1652

We still have some calls to gnutls functions where we do not check for
potential errors, but cleaning that up can be a separate patch.