build: Fix VPATH build when nbdkit not installed
Even when nbdkit is previously installed, we want to make sure we are
using the in-tree nbdkit-version.h, and not the pre-installed one.
Noticed when a native VPATH build succeeded but a mingw cross-build
failed with:

In file included from ../../../include/nbdkit-plugin.h:38,
                 from ../../../common/utils/environ.c:40:
../../../include/nbdkit-common.h:51:10: fatal error: nbdkit-version.h: No such file or directory
   51 | #include <nbdkit-version.h>
      |          ^~~~~~~~~~~~~~~~~~
compilation terminated.

We don't (currently) have a VPATH CI job, and even if we did, we may
not have spotted this as easily, as lcitool likes to prepopulate the
build containers with the distro's prebuilt nbdkit when possible.

At any rate, nbdkit-version.h is generated, so for a VPATH build to
find it, we must also include relative to builddir.

Fixes: 0c75ccde ("include: Expose nbdkit version information to public", v1.15.1)
(cherry picked from commit 8e7dc2af4ffd4154c8a45a39df99149daa506e6b)