build: Only attempt to build vddk on Linux
When --enable/disable-vddk is not given to configure, our default was
to base on the current architecture.  But we know that we are
targeting a .so library built for Linux, so we can also gate things
based on the host OS.  And doing so means that vddk is no longer even
attempted on mingw, eliminating the need to explicitly enable/disable
it in our CI recipes.