python: List function names in traceback
When using PyArg_ParseTuple, it is worth informing python of the
method name to list in tracebacks should the parse fail.  And for
parse_size, we are better letting the user get the usual error message
already produced by PyArg_ParseTuple than trying to substitute it with
one of our own on failure.