api: Add nbdkit_disconnect(int)
We have already exposed nbdkit_shutdown() as a way for a plugin to
request that nbdkit itself shuts down when convenient (taking away the
current and all other client connections).  But we lacked a way for a
plugin (or filter) to drop the current client, while still leaving
nbdkit up for other clients.  This has utility for testing client
reconnect logic, as well as for emulating things like qemu's
propensity to hard-disconnect on NBD_CMD_WRITE with more than 32M of
data (even though nbdkit would normally accept up to 64M of data).
Add a new API nbdkit_disconnect(int force) for that purpose, utilizing
the logic already built in to connections.c.  If force is true, we
close the socket immediately (the client won't get responses to the
current or any other parallel commands); if force is false, we mark
the socket as not accepting further commands (current commands can
still give their response to the client, but new requests from the
current client are ignored with ESHUTDOWN until the client gives
NBD_CMD_DISC).

This patch provides the C interface (using int, to allow for future
extensions beyond our current two values), and straightforward
bindings to languages that already have nbdkit_shutdown bound (here,
we use bool where it exists, since we don't have to worry as much
about potential future API breaks if we need a third state in the
future).  Upcoming patches will then work on extending the sh/eval
plugin to take advantage of the API, as well as enhancing the
blocksize-policy filter to more closely emulate qemu.

The logic for connection status changes slightly - there is now a
fourth state tracking when we want a soft shutdown, but are still
waiting for the client to send NBD_CMD_DISC - existing commands still
get to send their reply unmolested, but future commands get an
immediate ESHUTDOWN error.  Rather than re-checking the state between
dropping the read lock and grabbing the write lock, we now go by the
state present after recv() first gives us a result at the start of a
transaction.

The testsuite covers both hard and soft disconnects, and in both
plaintext and tls.

Message-Id: <20221026221802.207215-4-eblake@redhat.com>
Reviewed-by: Richard W.M. Jones <rjones@redhat.com>