sh: Advertise max known status in --dump-plugin
We are about to define specific meaning to more return status values.
To make it easier to probe if nbdkit is new enough to honor such
definitions (and thus whether it makes sense to write a script that
tries to utilize such special return values), it is worth advertising
the max known status supported by the given build of the sh/eval
plugin.

Message-Id: <20221109204353.1342583-2-eblake@redhat.com>
Reviewed-by: Laszlo Ersek <lersek@redhat.com>