sh: Add exit status triggers for nbdkit_{shutdown, disconnect}
Make it possible for the sh and eval plugins to disconnect a client or
shut down the entire nbdkit server by use of special return values.
Prior to this patch we had reserved 4-7 for future use; this defines
4-8, and extends the set of reserved return values to 9-15.  We figure
it is unlikely that anyone is using status 4-15 with the intent that
it behaves identically to status 1; more importantly, the choice of
soft disconnect with error for status 8 is the one least likely to
break such an existing sh or eval script.

For the testsuite, I only covered the eval plugin; but since it shares
common code with the sh plugin, both styles should work.

A note about the pod documentation: when the argument to =item would
otherwise appear to be a single number, the use of S<> to mask it is
necessary.  Newer perl documents the use of Z<> for this, but I didn't
want to risk failure of older perl not understanding that.

Message-Id: <20221109204353.1342583-3-eblake@redhat.com>
[split out refactoring, fix doc typos, further simplify switch]
Reviewed-by: Laszlo Ersek <lersek@redhat.com>