curl: Fix verification of CURLOPT_TLS13_CIPHERS
The code checking for CURLOPT_TLS13_CIPHERS option did not work
properly, because of incorrect assumption that this symbol was a
preprocessor macro. It is in fact element of enum type, which
resulted with #ifdef directive working improperly. Fix changes that
check to be based on curl version instead.

Signed-off-by: Michal Orzel <michalx.orzel@intel.com>
(cherry picked from commit 242757dd5c9fbf00a487ab934d67db442fe08661)