tests/test-blocksize-write-disconnect.sh: Consume input in pwrite
Scripts must always consume the input in pwrite calls, even if they
later return an error.  We also have to send the stderr output of the
dd command to /dev/null so it doesn't mess with the intended error
message ("ENOMEM").

This caused the following problem when running under valgrind
(probably not caused by valgrind, but by the slower nbdkit revealing a
race condition):

  nbdkit: eval[1]: debug: calling: /tmp/nbdkitdCGqA8/pwrite pwrite "" 8388610 0 ""
  nbdkit: eval[1]: error: /tmp/nbdkitdCGqA8/pwrite: write to script failed because of a broken pipe: this can happen if the script exits without consuming stdin, which usually indicates a bug in the script

We also get the following assert-fail (newly added in
commit 53672b859f "sh: Refactor handle_script_error()") but I will
address that separately:

  nbdkit: call.c:443: handle_script_error: Assertion `ebuf->ptr' failed.

Fixes: commit 6a7def7e259e87c01658a2164798baf06d98b99b