tests/test-readahead.sh: Remove bogus "$@"
In commit db1e331172 ("readahead: Fix test") I made this change
(edited for clarity):

  -test ()
  -{
  -    nbdkit -fv -U - "$@" null size=1M --filter=delay rdelay=5 \
  +nbdkit -fv -U - "$@" sh - \

In lifting the invocation of nbdkit from a function to the top level I
accidentally copied "$@" (variable function arguments).  This didn't
break the test because the script is called with no arguments so "$@"
expanded to nothing, but it was obviously wrong.

Fixes: commit db1e3311727c6ecab3264a1811d33db1aa45a4d0