luks: Avoid crash when image does not contain a LUKS header
We attempt to load the LUKS header in the prepare() callback.  If this
fails, h->h will be NULL and we'll crash in close() when we attempt to
access and free h->h->masterkey.

This crash could have been triggered another way: if open() followed
by close() was called, without prepare() or other callbacks.

Reported-by: Ming Xie
Fixes: https://bugzilla.redhat.com/show_bug.cgi?id=2159581