server/crypto.c: Use internal debug macro instead of nbdkit_debug function
The convention within the server code is to use the debug macro
instead of the nbdkit_debug function.  A few places in server/crypto.c
didn't do that.

The function and macro are not quite equivalent because the macro
checks if 'verbose' is set in the expanded code.  (The function checks
the flag again.)  The difference is not relevant here.

Updates: commit fab4a090d09fa213ee0b6fb8fbaa386ef9a6641c
(cherry picked from commit 2600bbaafd3833adc2bb9053956a54497864b86c)