server: Embolden debug messages from filters and plugins
It's easier to pick out debug messages from plugins against the
general noise of nbdkit debugging if they are displayed in a bold
font.

Route the debug() macro (ie. debug messages coming from the server)
through a new internal function debug_in_server().  The debug_common
function added in the previous commit gets a new boolean flag to
indicate whether we were called from the server or from somewhere else
(ie. a filter or plugin).  We can use this information, if stderr is
going to a terminal, to embolden filter and plugin messages.

Note the public API/ABI of nbdkit_debug and nbdkit_vdebug is not
changed.