curl: Fix ssl-version so it checks value during plugin configuration
Fix this parameter so that instead of storing the string and resolving
it to a CURL_SSLVERSION_* value when creating the handle, we do it
when configuring the plugin.  This will ensure that errors are
detected early (the plugin will not start up), rather than having the
plugin start up and then fail on first connection.

Before this change:

  $ nbdkit -f curl https://foo ssl-version=blah

(no error until you connect first time)

After this change:

  $ nbdkit -f curl https://foo ssl-version=blah
  nbdkit: error: unknown ssl-version: blah

Fixes: commit f692e1001db73b9aca29c521268812de73bdd0f2